refactor: UserTest 추상화
